Refactor: move styled-components logic to separate file
