+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Home from "./pages/Home";
+import Home from "./pages/Home/Home";
 import GlobalStyle from "../style/GlobalStyle";
 
 const App = () => {
