--- src/components/MapContainer.jsx
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-const MapContainer = () => {
-  return <Wrapper>MapContainer</Wrapper>;
-};
-
-export default MapContainer;
-
-const Wrapper = styled.div`
-  width: 100%;
-  background-color: var(--color-gray-light);
-`;
