+++ src/components/MapContainer/MapContainer.jsx
@@ -1,0 +1,7 @@
+import { Wrapper } from "./MapContainer.styled";
+
+const MapContainer = () => {
+  return <Wrapper>MapContainer</Wrapper>;
+};
+
+export default MapContainer;
