+++ src/components/MapContainer/MapContainer.styled.jsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+  width: 100%;
+  background-color: var(--color-gray-light);
+`;
