+++ src/components/SideBar/SideBar.styled.jsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.aside`
+  display: flex;
+  flex-direction: column;
+  width: 450px;
+  height: 100%;
+`;
