+++ src/pages/Home/Home.styled.jsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.main`
+  display: flex;
+  height: 100vh;
+`;
